Remove linting from the project
ESLint is overkill for this type of playground project and ESLint 9 which we use for the actual monaco integration is a pain to setup for TypeScript. And in addition keeping the versions and type definitions on separate versions for project linting and editor integration is a complete hassle that I don't need.